Add files via upload
Text Editor Web Application..
This is totally Based on Python and using the Tkinter Library..
cx_Freeze is used for creating the setup of this application where cx_Freeze is a set of script and modules for freezing Python Scripts into executable in much the same way that py2exe and py2app do. Unlike these two tools, cx_Freeze is cross platform and should work on any platform that python itself works on. 

Please remember its always support Python 3.5 or higher.